refactor to remove deps
